+++ certbot-compatibility-test/certbot_compatibility_test/validator_test.py
@@ -1,10 +1,7 @@
 """Tests for certbot_compatibility_test.validator."""
 import unittest
+from unittest import mock
 
-try:
-    import mock
-except ImportError: # pragma: no cover
-    from unittest import mock # type: ignore
 import OpenSSL
 import requests